requirements.txt
pytz
requests
python-dotenv~=1.0.0

ipython==8.5.0  # enhanced python interpreter

# Django
django==4.2.7
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
whitenoise==6.2.0  # for serving static files

# Django 3.0 async requirements
gunicorn==20.1.0
uvicorn==0.18.3

# Databases
psycopg2-binary
dj-database-url==1.0.0

# Telegram
python-telegram-bot==20.6

# monitoring
# sentry-sdk

# Dev dependencies
black
flake8
isort
pytest

# openai
openai==1.2.3
